refactor: move some class in Domain.Common to SharedKernel
