+++ src/GlobalUsing.cs
@@ -2,7 +2,11 @@ global using MediatR;
 global using Steve.ManagerHero.UserService.Application.DTO;
 global using Steve.ManagerHero.UserService.Application.Interfaces.Repository;
 global using Steve.ManagerHero.UserService.Domain.AggregatesModel;
-global using Steve.ManagerHero.SharedKernel.Application.Exception;
-global using Steve.ManagerHero.UserService.Domain.Common;
 global using Steve.ManagerHero.UserService.Domain.Exception;
+
+// Shared Kernel
+global using Steve.ManagerHero.SharedKernel.Application.BaseClasses;
+global using Steve.ManagerHero.SharedKernel.Application.Interface;
 global using Steve.ManagerHero.SharedKernel.Application.Constants;
+global using Steve.ManagerHero.SharedKernel.Application.Pagination;
+global using Steve.ManagerHero.SharedKernel.Application.Exception;
