+++ src/SharedKernel/Application/Interface/IUnitOfWork.cs
@@ -4,7 +4,7 @@
 * - [2025-04-11] - Created by mrsteve.bang@gmail.com
 */
 
-namespace Steve.ManagerHero.UserService.Domain.Common;
+namespace Steve.ManagerHero.SharedKernel.Application.Interface;
 
 /// <summary>
 /// Represents a unit of work.
